33:42 삭제 갱신처리 추가(동작확인 완료)

+++ boardMongo/src/main/java/com/example/boardMongo/BoardMongoController.java
@@ -85,27 +85,28 @@ public class BoardMongoController {
 	
 	@RequestMapping(value="/mod.do", method=RequestMethod.POST) 
     @ResponseBody
-	public Map<String, Object> mod(@RequestParam(value="id_hidden", required=true) String id, 
+	public Map<String, Object> mod(
+			@RequestParam(value="id", required=true) String id, 
 			@RequestParam(value="title", required=true) String title,
 			@RequestParam(value="contents", required=false,defaultValue="") String contents) throws Exception {
 		System.out.println("mod=========");
+		
 		Map<String, Object> map = new HashMap<>();
 				
 		try {
 			Query query = new Query();
 			Criteria activityCriteria = Criteria.where("id").is(id);
 			query.addCriteria(activityCriteria);
-			activityCriteria = Criteria.where("title").is(title);
-			query.addCriteria(activityCriteria);
+			//activityCriteria = Criteria.where("title").is(title);
+			//query.addCriteria(activityCriteria);
 			List<Board> out = mongoTemplate.find(query, Board.class);
 			
 			if(out.size() > 0) {
-				Board in= out.get(0);
+				Board in= out.get(0);	//한건 삭제
 				in.setTitle(title);
 				in.setContents(contents);
 				boardRepository.save(in);
 			}
-			
 			map.put("returnCode", "success");
 			map.put("returnDesc", "정상적으로 수정되었습니다.");
 		}
@@ -119,16 +120,25 @@ public class BoardMongoController {
 	
 	@RequestMapping(value="/del.do", method=RequestMethod.POST) 
     @ResponseBody
-	public Map<String, Object> del(@RequestParam(value="title", required=true) String title, 
+	public Map<String, Object> del(
+			@RequestParam(value="id", required=true) String id, 
+			@RequestParam(value="title", required=true) String title, 
 			@RequestParam(value="contents", required=false,defaultValue="") String contents) throws Exception {
 		System.out.println("del=========");
 		Map<String, Object> map = new HashMap<>();
 				
 		try {
-			Board in = new Board();
-			in.setTitle(title);
-			in.setContents(contents);
-			boardRepository.delete(in);
+			Query query = new Query();
+			Criteria activityCriteria = Criteria.where("id").is(id);
+			query.addCriteria(activityCriteria);
+			activityCriteria = Criteria.where("title").is(title);
+			query.addCriteria(activityCriteria);
+			List<Board> out = mongoTemplate.find(query, Board.class);
+			
+			if(out.size() > 0) {
+				Board in= out.get(0);	//한건 삭제
+				boardRepository.delete(in);
+			}
 			
 			map.put("returnCode", "success");
 			map.put("returnDesc", "정상적으로 삭제되었습니다.");
